Fix styling and crashes for selected stories block type
