+++ package.json
@@ -38,6 +38,7 @@
     "@web-stories-wp/activation-notice": "*",
     "@web-stories-wp/date": "*",
     "@web-stories-wp/fonts": "*",
+    "@web-stories-wp/glider": "*",
     "@web-stories-wp/i18n": "*",
     "@web-stories-wp/jest-amp": "*",
     "@web-stories-wp/jest-puppeteer-amp": "*",
