+++ packages/glider/README.md
@@ -1,0 +1,3 @@
+# Glider
+
+Patched version of [Glider.js](https://github.com/NickPiscitelli/Glider.js).
